refactor(backend): split supabase_db.py into domain-specific modules
Split the 1153-line supabase_db.py into focused modules:
- _base.py: shared SupabaseClientMixin with client init and error handling
- collections_db.py: CollectionsDB and collection CRUD operations
- documents_db.py: SupabaseVectorDB and legal document/vector search operations
- publications_db.py: PublicationsDB and publication management operations
- schemas_db.py: SupabaseDB and extraction schema CRUD operations
- supabase_db.py: backward-compatible re-export barrel (all public names preserved)

All existing imports from juddges_search.db.supabase_db continue to work
without any consumer changes.

Closes #72
